♻️ move pure serialisation and byte utilities to @datadog/js-core/util
Move byteUtils, context types, jsonStringify, objectValues and isServerError
to js-core — they have no browser-specific dependencies and are needed by the
transport modules that will follow in subsequent commits.

Browser-core files become thin re-exports so all existing relative imports
keep working without touching every consumer.
