+++ packages/browser-core/src/tools/serialisation/context.ts
@@ -1,11 +1,1 @@
-export interface Context {
-  [x: string]: ContextValue
-}
-
-export type ContextValue = string | number | boolean | Context | ContextArray | undefined | null
-
-/**
- * @hidden
- */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface ContextArray extends Array<ContextValue> {}
+export type { Context, ContextValue, ContextArray } from '@datadog/js-core/util'
