+++ packages/browser-core/src/tools/utils/polyfills.ts
@@ -1,3 +1,5 @@
+export { objectValues } from '@datadog/js-core/util'
+
 export function findLast<T, S extends T>(
   array: readonly T[],
   predicate: (item: T, index: number, array: readonly T[]) => item is S
@@ -13,10 +15,6 @@ export function findLast<T, S extends T>(
 
 // Keep the following wrapper functions as it can be mangled and will result in smaller bundle size that using
 // the native Object.values and Object.entries directly
-
-export function objectValues<T = unknown>(object: { [key: string]: T }) {
-  return Object.values(object)
-}
 
 export function objectEntries<T = unknown>(object: { [key: string]: T }): Array<[string, T]> {
   return Object.entries(object)
